# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal AI Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gemini/conversations:
    get:
      operationId: listGeminiConversations
      tags: [gemini]
      summary: List all conversations for the authenticated user
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiConversation"
        "401":
          description: Unauthorized
    post:
      operationId: createGeminiConversation
      tags: [gemini]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversation"
        "401":
          description: Unauthorized
  /gemini/conversations/{id}:
    get:
      operationId: getGeminiConversation
      tags: [gemini]
      summary: Get conversation by ID with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversationWithMessages"
        "404":
          description: Not found
    patch:
      operationId: updateGeminiConversation
      tags: [gemini]
      summary: Update a conversation (title, pinned)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiConversationUpdate"
      responses:
        "200":
          description: Updated conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversation"
        "404":
          description: Not found
    delete:
      operationId: deleteGeminiConversation
      tags: [gemini]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /gemini/conversations/{id}/messages:
    get:
      operationId: listGeminiMessages
      tags: [gemini]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiMessage"
    post:
      operationId: sendGeminiMessage
      tags: [gemini]
      summary: Send a message and receive streaming AI response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiMessageInput"
      responses:
        "200":
          description: Streaming response (text/event-stream)
          content:
            text/event-stream: {}
  /gemini/conversations/{id}/export:
    get:
      operationId: exportConversation
      tags: [gemini]
      summary: Export conversation as Markdown
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Markdown export
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationExport"
        "404":
          description: Not found
  /gemini/generate-image:
    post:
      operationId: generateGeminiImage
      tags: [gemini]
      summary: Generate an image from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiImageInput"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiImageOutput"
...
  /gemini/digest:
    get:
      operationId: getAiDigest
      tags: [gemini]
      summary: Get AI-generated daily digest from recent conversations
      responses:
        "200":
          description: Daily digest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiDigest"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    GeminiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        pinned:
          type: boolean
        messageCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - pinned
        - messageCount
        - createdAt
        - updatedAt
    GeminiConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    GeminiConversationUpdate:
      type: object
      properties:
        title:
          type: string
        pinned:
          type: boolean
    GeminiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        imageData:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    GeminiMessageInput:
      type: object
      properties:
        content:
          type: string
        model:
          type: string
        systemPrompt:
          type: string
        imageBase64:
          type: string
        imageMimeType:
          type: string
        reasoningMode:
          type: boolean
      required:
        - content
    GeminiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        pinned:
          type: boolean
        messageCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/GeminiMessage"
      required:
        - id
        - title
        - pinned
        - messageCount
        - createdAt
        - updatedAt
        - messages
    GeminiImageInput:
      type: object
      properties:
        prompt:
          type: string
        style:
          type: string
        aspectRatio:
          type: string
      required:
        - prompt
    GeminiImageOutput:
      type: object
      properties:
        b64_json:
          type: string
        mimeType:
          type: string
      required:
        - b64_json
        - mimeType
...
    ConversationExport:
      type: object
      properties:
        title:
          type: string
        markdown:
          type: string
        exportedAt:
          type: string
          format: date-time
      required:
        - title
        - markdown
        - exportedAt
    AiDigest:
      type: object
      properties:
        summary:
          type: string
        generatedAt:
          type: string
          format: date-time
        conversationCount:
          type: integer
      required:
        - summary
        - generatedAt
        - conversationCount
...